+++ flows/actions/base.go
@@ -15,6 +15,7 @@ import (
 	"github.com/nyaruka/goflow/excellent/types"
 	"github.com/nyaruka/goflow/flows"
 	"github.com/nyaruka/goflow/flows/events"
+	"github.com/nyaruka/goflow/flows/modifiers"
 	"github.com/nyaruka/goflow/utils"
 
 	"github.com/pkg/errors"
@@ -162,7 +163,7 @@ func (a *baseAction) updateWebhook(run flows.Run, call *flows.WebhookCall) {
 
 // helper to apply a contact modifier
 func (a *baseAction) applyModifier(run flows.Run, mod flows.Modifier, logModifier flows.ModifierCallback, logEvent flows.EventCallback) {
-	mod.Apply(run.Environment(), run.Session().Assets(), run.Contact(), logEvent)
+	modifiers.Apply(run.Environment(), run.Session().Engine().Services(), run.Session().Assets(), run.Contact(), mod, logEvent)
 	logModifier(mod)
 }