+++ flows/actions/call_resthook.go
@@ -115,13 +115,13 @@ func (a *CallResthookAction) Execute(run flows.Run, step flows.Step, logModifier
 
 		req.Header.Add("Content-Type", "application/json")
 
-		svc, err := run.Session().Engine().Services().Webhook(run.Session())
+		svc, err := run.Session().Engine().Services().Webhook(run.Session().Assets())
 		if err != nil {
 			logEvent(events.NewError(err))
 			return nil
 		}
 
-		call, err := svc.Call(run.Session(), req)
+		call, err := svc.Call(req)
 
 		if err != nil {
 			logEvent(events.NewError(err))