+++ flows/actions/call_webhook.go
@@ -135,13 +135,13 @@ func (a *CallWebhookAction) call(run flows.Run, step flows.Step, url, method, bo
 		req.Header.Add(key, headerValue)
 	}
 
-	svc, err := run.Session().Engine().Services().Webhook(run.Session())
+	svc, err := run.Session().Engine().Services().Webhook(run.Session().Assets())
 	if err != nil {
 		logEvent(events.NewError(err))
 		return nil
 	}
 
-	call, err := svc.Call(run.Session(), req)
+	call, err := svc.Call(req)
 
 	if err != nil {
 		logEvent(events.NewError(err))