+++ flows/actions/send_email.go
@@ -100,13 +100,13 @@ func (a *SendEmailAction) Execute(run flows.Run, step flows.Step, logModifier fl
 		return nil
 	}
 
-	svc, err := run.Session().Engine().Services().Email(run.Session())
+	svc, err := run.Session().Engine().Services().Email(run.Session().Assets())
 	if err != nil {
 		logEvent(events.NewError(err))
 		return nil
 	}
 
-	err = svc.Send(run.Session(), evaluatedAddresses, evaluatedSubject, evaluatedBody)
+	err = svc.Send(evaluatedAddresses, evaluatedSubject, evaluatedBody)
 	if err != nil {
 		logEvent(events.NewError(errors.Wrap(err, "unable to send email")))
 	} else {