+++ flows/engine/engine_test.go
@@ -22,9 +22,9 @@ func TestBuilder(t *testing.T) {
 	assert.EqualError(t, err, "no email service factory configured")
 	_, err = eng.Services().Airtime(nil)
 	assert.EqualError(t, err, "no airtime service factory configured")
-	_, err = eng.Services().Classification(nil, nil)
+	_, err = eng.Services().Classification(nil)
 	assert.EqualError(t, err, "no classification service factory configured")
-	_, err = eng.Services().Ticket(nil, nil)
+	_, err = eng.Services().Ticket(nil)
 	assert.EqualError(t, err, "no ticket service factory configured")
 	_, err = eng.Services().Webhook(nil)
 	assert.EqualError(t, err, "no webhook service factory configured")
@@ -33,7 +33,7 @@ func TestBuilder(t *testing.T) {
 	webhookSvc := webhooks.NewService(&http.Client{}, nil, nil, map[string]string{"User-Agent": "goflow"}, 1000)
 
 	eng = engine.NewBuilder().
-		WithWebhookServiceFactory(func(flows.Session) (flows.WebhookService, error) { return webhookSvc, nil }).
+		WithWebhookServiceFactory(func(flows.SessionAssets) (flows.WebhookService, error) { return webhookSvc, nil }).
 		Build()
 
 	svc, err := eng.Services().Webhook(nil)