+++ services/email/smtp/service_test.go
@@ -23,13 +23,13 @@ func TestService(t *testing.T) {
 	svc, err := smtp.NewService("smtp://leah:pass123@temba.io:255?from=updates@temba.io", nil)
 	require.NoError(t, err)
 
-	err = svc.Send(nil, []string{"bob@nyaruka.com", "jim@nyaruka.com"}, "Updates", "Have a great week")
+	err = svc.Send([]string{"bob@nyaruka.com", "jim@nyaruka.com"}, "Updates", "Have a great week")
 
 	assert.NoError(t, err)
 	assert.Equal(t, []string{"HELO localhost\nMAIL FROM:<updates@temba.io>\nRCPT TO:<bob@nyaruka.com>\nRCPT TO:<jim@nyaruka.com>\nDATA\nHave a great week\n.\nQUIT\n"}, sender.Logs())
 
 	// if body is blank, we'll use a placeholder
-	err = svc.Send(nil, []string{"bob@nyaruka.com", "jim@nyaruka.com"}, "Updates", " ")
+	err = svc.Send([]string{"bob@nyaruka.com", "jim@nyaruka.com"}, "Updates", " ")
 
 	assert.NoError(t, err)
 	assert.Equal(t, "HELO localhost\nMAIL FROM:<updates@temba.io>\nRCPT TO:<bob@nyaruka.com>\nRCPT TO:<jim@nyaruka.com>\nDATA\n(empty body)\n.\nQUIT\n", sender.Logs()[1])