+++ services/webhooks/service_test.go
@@ -158,8 +158,8 @@ func TestWebhookParsing(t *testing.T) {
 		request, err := http.NewRequest(tc.call.method, tc.call.url, strings.NewReader(tc.call.body))
 		require.NoError(t, err)
 
-		svc, _ := session.Engine().Services().Webhook(session)
-		c, err := svc.Call(session, request)
+		svc, _ := session.Engine().Services().Webhook(session.Assets())
+		c, err := svc.Call(request)
 
 		if tc.isError {
 			assert.Error(t, err, "expected error for call %s", tc.call)
@@ -192,8 +192,8 @@ func TestRetries(t *testing.T) {
 	request, err := http.NewRequest("GET", "http://temba.io/", strings.NewReader("BODY"))
 	require.NoError(t, err)
 
-	svc, _ := session.Engine().Services().Webhook(session)
-	c, err := svc.Call(session, request)
+	svc, _ := session.Engine().Services().Webhook(session.Assets())
+	c, err := svc.Call(request)
 	require.NoError(t, err)
 
 	assert.Equal(t, 200, c.Response.StatusCode)
@@ -211,7 +211,7 @@ func TestAccessRestrictions(t *testing.T) {
 	assert.NoError(t, err)
 
 	request, _ := http.NewRequest("GET", "http://localhost/foo", nil)
-	call, err := svc.Call(nil, request)
+	call, err := svc.Call(request)
 
 	// actual error becomes a call with a connection error
 	assert.NoError(t, err)
@@ -235,8 +235,8 @@ func TestGzipEncoding(t *testing.T) {
 
 	request.Header.Set("Accept-Encoding", "gzip")
 
-	svc, _ := session.Engine().Services().Webhook(session)
-	c, err := svc.Call(session, request)
+	svc, _ := session.Engine().Services().Webhook(session.Assets())
+	c, err := svc.Call(request)
 	require.NoError(t, err)
 
 	// check that gzip decompression happens transparently